* #8542: test for .git existence before moving pre-commit hook

+++ daemon/autogen.sh
@@ -5,10 +5,13 @@ mkdir -p m4
 
 HOOKS_DIR="../.git/hooks"
 # install pre-commit hook for doing clean commits
-if test ! \( -x ${HOOKS_DIR}/pre-commit -a -L ${HOOKS_DIR}/pre-commit \);
+if [ -d "$HOOKS_DIR" ];
 then
-    rm -f ${HOOKS_DIR}/pre-commit
-    ln -s ${HOOKS_DIR}/pre-commit.sample ${HOOKS_DIR}/pre-commit
+    if test ! \( -x ${HOOKS_DIR}/pre-commit -a -L ${HOOKS_DIR}/pre-commit \);
+    then
+        rm -f ${HOOKS_DIR}/pre-commit
+        ln -s ${HOOKS_DIR}/pre-commit.sample ${HOOKS_DIR}/pre-commit
+    fi
 fi
 
 autoreconf --force --install --verbose -Wall -I m4
